model-training-service/src/two-two-training/requirements.txt
torch>=2.0.0
torchvision
pandas
pyarrow
pyyaml
mlflow
scikit-learn
kubeflow-sdk